Move TrustSummaryDto to a Dto namespace
There's going to be a lot of these so tidying up into a specific directory
